+++ src/main/java/com/clover/habbittracker/domain/member/service/MemberService.java
@@ -2,10 +2,8 @@ package com.clover.habbittracker.domain.member.service;
 
 import com.clover.habbittracker.domain.member.dto.MemberRequest;
 import com.clover.habbittracker.domain.member.dto.MemberResponse;
-import com.clover.habbittracker.global.security.oauth.dto.SocialUser;
 
 public interface MemberService {
-	Long join(SocialUser socialUser);
 	MemberResponse getProfile(Long memberId);
 
 	MemberResponse updateProfile(Long memberId, MemberRequest request);
